Add one to ranges intended for insert:
Inserting is done by pointing to the element before which we want
to insert. But by inverting, this becomes the element after we want
to insert. Thus, we have to add one to the range.

Inverting the direction is surprisingly tricky! Perhaps we should
have simply inverted the source model?

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>