MobileSwipeModel: fix insertion of dives
Only the topLevel case was correcting for the fact that by
reversing the direction, we have to add 1 to the index.
This is because if we point to the item where *before* which
we want to insert in the source model, that is the item *after*
we insert in the swipe model.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>